requirements.txt
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
asgiref==3.8.1
babel==2.16.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.10
django-cleanup==9.0.0
django-cors-headers==4.3.1
django-filter==24.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
fake-http-header==0.3.5
filelock==3.16.1
ghp-import==2.1.0
gitignore_parser==0.1.11
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.0.0
Jinja2==3.1.6
jiter==0.8.2
Markdown==3.7
markdownify==0.14.1
MarkupSafe==3.0.2
mergedeep==1.3.4
mike==2.1.3
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.44
mkdocs-material-extensions==1.3.1
mockito==1.5.3
openai==1.69.0
packaging==24.2
paginate==0.5.7
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
playwright==1.49.1
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
pyee==12.0.0
Pygments==2.18.0
pymdown-extensions==10.12
pyparsing==3.2.3
pyproject-api==1.8.0
pytest==8.3.4
pytest-base-url==2.1.0
pytest-mockito==0.0.4
pytest-playwright==0.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
pyyaml_env_tag==0.1
regex==2024.11.6
requests==2.32.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
text-unidecode==1.3
tf-playwright-stealth==1.1.2
tox==4.23.2
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.5.0
uWSGI==2.0.28
verspec==0.1.0
virtualenv==20.27.1
watchdog==6.0.0
zipp==3.21.0
ruff==0.11.7
